# Use Plate Recognizer instead of OpenALPR