--- a/package.json
+++ b/package.json
@@ -58,7 +58,6 @@
     "normalize.css": "^8.0.0",
     "object-to-formdata": "^1.4.3",
     "object.omit": "^3.0.0",
-    "openalpr_api": "^2.0.1",
     "p-event": "^1.3.0",
     "parse": "^1.11.1",
     "polygon-lookup": "^2.4.0",