--- a/src/routes/home/Home.js
+++ b/src/routes/home/Home.js
@@ -604,7 +604,7 @@ class Home extends React.Component {
       formData.append('attachmentFile', attachmentBlob);
       const { data } = await axios.post('/openalpr', formData);
       const result = data.results[0];
-      result.licenseState = result.region.toUpperCase();
+      result.licenseState = result.region.code.split('-')[1].toUpperCase();
       if (
         this.state.plate === '' &&
         document.activeElement !== this.plateRef.current