--- a/src/server.js
+++ b/src/server.js
@@ -18,7 +18,6 @@ import nodeFetch from 'node-fetch';
 import React from 'react';
 import ReactDOM from 'react-dom/server';
 import PrettyError from 'pretty-error';
-import OpenalprApi from 'openalpr_api';
 import Parse from 'parse/node';
 import FileType from 'file-type/browser';
 import sharp from 'sharp';
@@ -27,6 +26,7 @@ import multer from 'multer';
 import stringify from 'json-stringify-safe';
 import DelayedResponse from 'http-delayed-response';
 import { JSDOM } from 'jsdom';
+import FormData from 'form-data';
 
 import { isImage, isVideo } from './isImage.js';
 import { validateLocation, processValidation } from './geoclient.js';
@@ -57,7 +57,7 @@ const {
   PARSE_MASTER_KEY,
   PARSE_SERVER_URL,
   HEROKU_RELEASE_VERSION,
-  OPENALPR_SECRET_KEY,
+  PLATERECOGNIZER_TOKEN,
 } = process.env;
 
 require('heroku-self-ping')(config.api.serverUrl, {
@@ -487,42 +487,32 @@ app.use('/submit', (req, res) => {
   });
 });
 
-// adapted from https://github.com/openalpr/cloudapi/tree/8141c1ba57f03df4f53430c6e5e389b39714d0e0/javascript#getting-started
-app.use('/openalpr', upload.single('attachmentFile'), (req, res) => {
-  const country = 'us';
-  const opts = {
-    recognizeVehicle: 1,
-    state: 'ny',
-    returnImage: 0,
-    topn: 10,
-    prewarp: '',
-  };
-
+// adapted from https://docs.platerecognizer.com/?javascript#license-plate-recognition
+app.use('/platerecognizer', upload.single('attachmentFile'), (req, res) => {
   const attachmentBuffer = req.file.buffer;
-  const api = new OpenalprApi.DefaultApi();
-
-  const secretKey = OPENALPR_SECRET_KEY; // {String} The secret key used to authenticate your account. You can view your secret key by visiting https://cloud.openalpr.com/
 
   orientImageBuffer({ attachmentBuffer })
     .then(buffer => buffer.toString('base64'))
     .then(attachmentBytesRotated => {
-      console.time(`/openalpr recognizeBytes`); // eslint-disable-line no-console
-      return new Promise((resolve, reject) => {
-        api.recognizeBytes(
-          attachmentBytesRotated,
-          secretKey,
-          country,
-          opts,
-          (error, data) => {
-            console.timeEnd(`/openalpr recognizeBytes`); // eslint-disable-line no-console
-            if (error) {
-              reject(error);
-            } else {
-              resolve(data);
-            }
-          },
-        );
-      });
+      console.log('STARTING platerecognizer'); // eslint-disable-line no-console
+      console.time(`/platerecognizer plate-reader`); // eslint-disable-line no-console
+
+      const body = new FormData();
+
+      body.append('upload', attachmentBytesRotated);
+
+      // body.append("regions", "us-ny"); // Change to your country
+      body.append('regions', 'us'); // Change to your country
+
+      return nodeFetch('https://api.platerecognizer.com/v1/plate-reader/', {
+        method: 'POST',
+        headers: {
+          Authorization: `Token ${PLATERECOGNIZER_TOKEN}`,
+        },
+        body,
+      })
+        .then(platerecognizerRes => platerecognizerRes.json())
+        .finally(() => console.timeEnd(`/platerecognizer plate-reader`)); // eslint-disable-line no-console
     })
     .then(data => res.json(data))
     .catch(handlePromiseRejection(res));