The OpenALPR trial we're using expires on 12 July 2022,
so let's see if we can switch to a different API.

More context at https://reportedcab.slack.com/archives/C7Z0BHMMY/p1656542848126409?thread_ts=1656276082.804279&cid=C7Z0BHMMY