@josephfrazier 2022-07-11
(rough performance notes from commit [b63dae5](https://github.com/josephfrazier/Reported-Web/pull/335/commits/b63dae5fb0d1c4b1e1f237f39a0aedd14874cbec))

before browser:

```
blobUtil.blobToArrayBuffer(attachmentFile): 7.941162109375 ms
Buffer.from(attachmentArrayBuffer): 0.108154296875 ms
bufferToBlob(PXL_20220707_135731889.jpg): 4.928955078125 ms
extractPlate: 3036.756103515625 ms
```

before server:

```
orientImageBuffer: 248.912ms
/openalpr recognizeBytes: 2.741s
```

after client:

```
blobUtil.blobToArrayBuffer(attachmentFile): 8.02099609375 ms
Buffer.from(attachmentArrayBuffer): 0.10205078125 ms
bufferToBlob(PXL_20220707_135731889.jpg): 5.114990234375 ms
extractPlate: 1222.174072265625 ms
```

after server:

```
image buffer length BEFORE sharp: 1299123 bytes
image buffer length AFTER sharp: 1587330 bytes
orientImageBuffer: 243.055ms
STARTING platerecognizer
/platerecognizer plate-reader: 928.728ms
```

image used:

<details>

![PXL_20220707_135731889](https://user-images.githubusercontent.com/6473925/178374863-a47c3b8f-e3c7-4342-9592-d2b492490c6b.jpg)

</details>